+++ LoveBird/Sources/Root/Onboarding/OnboardingNicknameView.swift
@@ -12,8 +12,7 @@ struct OnboardingNicknameView: View {
   
   let store: StoreOf<OnboardingCore>
   @FocusState private var isFocused: Bool
-  @StateObject private var keyboard = KeyboardResponder()
-  
+
   var body: some View {
     WithViewStore(self.store, observe: { $0 }) { viewStore in
       VStack(spacing: 10) {
@@ -40,7 +39,7 @@ struct OnboardingNicknameView: View {
           viewStore.send(.nextTapped)
         }
         .padding(.horizontal, 16)
-        .padding(.bottom, keyboard.currentHeight == 0 ? 20 + UIApplication.edgeInsets.bottom : keyboard.currentHeight + 20)
+        .padding(.bottom, 20)
       }
       .frame(maxWidth: .infinity, maxHeight: .infinity)
       .background(.white)
